create controller concern rdf_response_formats with methods refactored out of annotations_controller
