+++ app/controllers/triannon/annotations_controller.rb
@@ -2,6 +2,8 @@ require_dependency "triannon/application_controller"
 
 module Triannon
   class AnnotationsController < ApplicationController
+    include RdfResponseFormats
+
     before_action :default_format_jsonld, only: [:show]
     before_action :set_annotation, only: [:show, :update, :destroy]
     rescue_from Triannon::ExternalReferenceError, with: :ext_ref_error
@@ -51,7 +53,7 @@ module Triannon
     end
 
     # NOT YET IMPLEMENTED
-    # GET /annotations/1/edit    
+    # GET /annotations/1/edit
 #    def edit
 #    end
 
@@ -71,7 +73,7 @@ module Triannon
         content_type = request.headers["Content-Type"]
         @annotation = Annotation.new({:data => request.body.read, :expected_content_type => content_type})
       end
-      
+
       if @annotation.save
         default_format_jsonld # NOTE: this must be here and not in before_filter or we get Missing template errors
         flash[:notice] = "Annotation #{@annotation.id} was successfully created."
@@ -124,83 +126,18 @@ module Triannon
       @annotation.destroy
       redirect_to annotations_url, status: 204, notice: 'Annotation was successfully destroyed.'
     end
-    
+
 private
 
     def set_annotation
       @annotation = Annotation.find(params[:id])
-    end
-    
-    # set format to jsonld if it isn't already set
-    def default_format_jsonld
-      if ((!request.accept || request.accept.empty?) && (!params[:format] || params[:format].empty?))
-        request.format = "jsonld"
-      end
-    end
-
-    # find first mime type from request.accept that matches return mime type
-    def mime_type_from_accept(return_mime_types)
-      @mime_type_from_accept ||= begin
-        if request.accept && request.accept.is_a?(String)
-          accept_mime_types = request.accept.split(',')
-          accept_mime_types.each { |mime_type|
-            mime_str = mime_type.split("; profile=").first.strip
-            if return_mime_types.include? mime_str
-              return mime_str
-            end
-          }
-        end
-      end
-    end
-
-    # parse the Accept HTTP header for the value of profile if it is a request for jsonld or json
-    # e.g. Accept: application/ld+json; profile="http://www.w3.org/ns/oa-context-20130208.json"
-    # @return [String] url for jsonld @context or nil if missing or non-jsonld/json format
-    def context_url_from_accept
-      if request.format == "jsonld" || request.format == "json"
-        accept_str = request.accept
-        if accept_str && accept_str.split("profile=") && accept_str.split("profile=").last
-          context_url = accept_str.split("profile=").last.strip
-          context_url = context_url[1, context_url.size] if context_url.start_with?('"')
-          context_url = context_url[0, context_url.size-1] if context_url.end_with?('"')
-          case context_url
-            when Triannon::JsonldContext::OA_DATED_CONTEXT_URL, 
-              Triannon::JsonldContext::OA_CONTEXT_URL,
-              Triannon::JsonldContext::IIIF_CONTEXT_URL
-              context_url
-            else
-              nil
-          end
-        end
-      end
-    end
-
-    # parse the Accept HTTP Link for the value of rel if it is a request for jsonld or json
-    # e.g. Link: http://www.w3.org/ns/oa.json; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
-    # note that the "type" part is optional
-    # @return [String] url for jsonld @context or nil if missing or non-jsonld/json format
-    def context_url_from_link
-      if request.format == "jsonld" || request.format == "json"
-        link_str = request.headers["Link"]
-        if link_str && link_str.split("; rel=") && link_str.split("; rel=").first
-          context_url = link_str.split("; rel=").first.strip
-          case context_url
-            when Triannon::JsonldContext::OA_DATED_CONTEXT_URL,
-              Triannon::JsonldContext::OA_CONTEXT_URL,
-              Triannon::JsonldContext::IIIF_CONTEXT_URL
-              context_url
-            else
-              nil
-          end
-        end
-      end
     end
 
     # handle Triannon::ExternalReferenceError
     def ext_ref_error(exception)
       render plain: exception.message, status: 403
     end
-    
+
     # render json_ld respecting requested context
     # @param [String] req_context set to "iiif" or "oa".  Default is oa
     # @param [String] mime_type the mime type to be set in the Content-Type header of the HTTP response
